+++ models/itoyokado/itoyokado.go
@@ -1,7 +1,7 @@
 package itoyokado
 
 import (
-	"github.com/jinzhu/gorm"
+	"toriniku/models/common"
 )
 
 const (
@@ -17,11 +17,15 @@ const (
 )
 
 // Group 店舗テーブル
+// type Group struct {
+// 	ID         uint `gorm:"primary_key"`
+// 	ShopName   string
+// 	URL        string
+// 	Prefecture string
+// }
+
 type Group struct {
-	ID         uint `gorm:"primary_key"`
-	ShopName   string
-	URL        string
-	Prefecture string
+	common.Group
 }
 
 // TableName 店舗テーブル名
@@ -30,11 +34,15 @@ func (g Group) TableName() string {
 }
 
 // Stock 在庫テーブル
+// type Stock struct {
+// 	gorm.Model
+// 	ShopID      uint
+// 	ProductID   uint
+// 	ProductName string
+// }
+
 type Stock struct {
-	gorm.Model
-	ShopID      uint
-	ProductID   uint
-	ProductName string
+	common.Stock
 }
 
 // TableName 在庫テーブル名
@@ -43,12 +51,16 @@ func (s Stock) TableName() string {
 }
 
 // Product 商品テーブル
+// type Product struct {
+// 	gorm.Model
+// 	ShopName string
+// 	Product  string
+// 	Price    int
+// 	Per100G  int
+// }
+
 type Product struct {
-	gorm.Model
-	ShopName string
-	Product  string
-	Price    int
-	Per100G  int
+	common.Product
 }
 
 // TableName 在庫テーブル名
