+++ templates/index.html
@@ -92,22 +92,52 @@
     </h4>
     {{range .products}}
        <h5>
-       <a href="https://www.iy-net.jp/nspc/shoptop.do?shopcd=00239">ショップ名:{{.ShopName}}</a>
+       <a href="https://www.iy-net.jp/nspc/shoptop.do?shopcd=00239">ショップ名:{{.GroupName}}_{{.ShopName}}</a>
        </h5>
-      <h5>商品名:{{.Product}}</h5>
+      <h5>商品名:{{.Name}}</h5>
       <h5>100g当たりの最安値:{{.Per100G}}円</h5>
     {{end}}
   
-    <form method="POST" action="/aeon/shoplist" >
-      <input class="btn btn-info" type="submit" value="イトーヨーカドー 店舗一覧取得">
-    </form>
+    <div style="display:inline-flex">
+      <form method="POST" action="/itoyokado/shoplist" >
+        <input class="btn btn-info" type="submit" value="イトーヨーカドー 店舗一覧取得">
+      </form>
+  
+      <form method="POST" action="/itoyokado/search" >
+        <div>店舗コード：<input type="number" name="shopcode"></div>
+        <input class="btn btn-info" type="submit" value="イトーヨーカドーの指定された店舗から情報取得">
+      </form>    
+    </div>
 
-    <form method="POST" action="/aeon/search" >
-      <div>店舗コード：<input type="number" name="shopcode"></div>
-      <input class="btn btn-info" type="submit" value="指定された店舗から情報取得">
-    </form>
+    <br>
 
-    <form method="POST" action="/life/compare" >
+    <div style="display:inline-flex">
+      <form method="POST" action="/life/shoplist" >
+        <input class="btn btn-info" type="submit" value="ライフ 店舗一覧取得">
+      </form>
+  
+      <form method="POST" action="/life/search" >
+        <div>店舗コード：<input type="number" name="shopcode"></div>
+        <input class="btn btn-info" type="submit" value="ライフの指定された店舗から情報取得">
+      </form>
+    </div>
+
+    <br>
+
+    <div style="display:inline-flex">
+      <form method="POST" action="/aeon/shoplist" >
+        <input class="btn btn-info" type="submit" value="イオン店舗一覧取得">
+      </form>
+
+      <form method="POST" action="/aeon/search" >
+        <div>店舗コード：<input type="number" name="shopcode"></div>
+        <input class="btn btn-info" type="submit" value="イオンの指定された店舗から情報取得">
+      </form>      
+    </div>
+
+    <br>
+
+    <form method="POST" action="/common/compare" >
       <input class="btn btn-info" type="submit" value="とり肉の最新最安値">
     </form>
